[MFLOWPLAY-112] - StaticAnalysis errors fix - Moved public model classes outside internal package - Updated changelog - version = 1.3.0 (15)
