--- Changelogs/1.2.6
@@ -1,8 +1,0 @@
-1.2.6 Release notes (2023-11-02)
-=============================================================
-
-Improvements to the 'RingPublishingTracking' module.
-
-### Features
-
-* Added video events tracking
