+++ RingPublishingTracking/src/main/java/com/ringpublishing/tracking/data/video/VideoContentCategory.kt
@@ -1,4 +1,4 @@
-package com.ringpublishing.tracking.internal.video
+package com.ringpublishing.tracking.data.video
 
 /**
  * Video content category related params
@@ -20,5 +20,4 @@ enum class VideoContentCategory(val text: String) {
      * Video content was purchased using premium subscription
      */
     VOID_SUBSCRIPTION("svod")
-
 }
