+++ RingPublishingTracking/src/main/java/com/ringpublishing/tracking/data/video/VideoMetadata.kt
@@ -1,4 +1,4 @@
-package com.ringpublishing.tracking.internal.video
+package com.ringpublishing.tracking.data.video
 
 /**
  * Video meta data holder
