+++ RingPublishingTracking/src/main/java/com/ringpublishing/tracking/data/video/VideoState.kt
@@ -1,4 +1,4 @@
-package com.ringpublishing.tracking.internal.video
+package com.ringpublishing.tracking.data.video
 
 /**
  * Video state data holder
