+++ RingPublishingTracking/src/main/java/com/ringpublishing/tracking/data/video/VideoStreamFormat.kt
@@ -1,4 +1,4 @@
-package com.ringpublishing.tracking.internal.video
+package com.ringpublishing.tracking.data.video
 
 /**
  * Video stream format related params
