+++ RingPublishingTracking/src/test/java/com/ringpublishing/tracking/internal/factory/VideoEventsFactoryTest.kt
@@ -9,16 +9,16 @@ package com.ringpublishing.tracking.com.ringpublishing.tracking.internal.factory
 import android.util.Base64
 import com.google.gson.GsonBuilder
 import com.ringpublishing.tracking.internal.factory.EventType
-import com.ringpublishing.tracking.internal.factory.VideoEventParam
+import com.ringpublishing.tracking.internal.video.VideoEventParam
 import com.ringpublishing.tracking.internal.factory.VideoEventsFactory
-import com.ringpublishing.tracking.internal.video.VideoAdsConfiguration
-import com.ringpublishing.tracking.internal.video.VideoContentCategory
-import com.ringpublishing.tracking.internal.video.VideoEvent
-import com.ringpublishing.tracking.internal.video.VideoMetadata
-import com.ringpublishing.tracking.internal.video.VideoStartMode
-import com.ringpublishing.tracking.internal.video.VideoState
-import com.ringpublishing.tracking.internal.video.VideoStreamFormat
-import com.ringpublishing.tracking.internal.video.VideoVisibilityState
+import com.ringpublishing.tracking.data.video.VideoAdsConfiguration
+import com.ringpublishing.tracking.data.video.VideoContentCategory
+import com.ringpublishing.tracking.data.video.VideoEvent
+import com.ringpublishing.tracking.data.video.VideoMetadata
+import com.ringpublishing.tracking.data.video.VideoStartMode
+import com.ringpublishing.tracking.data.video.VideoState
+import com.ringpublishing.tracking.data.video.VideoStreamFormat
+import com.ringpublishing.tracking.data.video.VideoVisibilityState
 import io.mockk.every
 import io.mockk.mockkStatic
 import io.mockk.slot
