+++ gradle.properties
@@ -8,6 +8,6 @@ android.useAndroidX=true
 android.enableJetifier=true
 
 #Library version name
-sdk_version_name=1.2.6
+sdk_version_name=1.3.0
 #Library version code
 sdk_version_code=15
